Написал тест для санки офферов
