--- src/services/thunk/thunk-test/fech-offers.test.ts
@@ -1,37 +1,0 @@
-import {createApi} from '../../api';
-import MockAdapter from 'axios-mock-adapter';
-import thunk from 'redux-thunk';
-import {Action} from 'redux';
-import {State} from '../../../types/type-store';
-import {ThunkDispatch} from 'redux-thunk';
-import {configureMockStore} from '@jedmao/redux-mock-store';
-import {ApiRoute} from '../../../const';
-import {checkAuthAction} from '../check-auth-action';
-
-type AppThunkDispatch = ThunkDispatch<State, ReturnType<typeof createApi>, Action>
-
-const extractActionsTypes = (actions: Action<string>[]) => actions.map(({type}) => type);
-
-describe('Async actions', () => {
-  const axios = createApi();
-  const mockAxiosAdapter = new MockAdapter(axios);
-  const middleware = [thunk.withExtraArgument(axios)];
-  const mockStoreCreator = configureMockStore<State, Action<string>, AppThunkDispatch>(middleware);
-  let store: ReturnType<typeof mockStoreCreator>;
-
-  beforeEach(() => {
-    store = mockStoreCreator({});
-  });
-
-  it('should dispatch "checkAuthAction.pending" and "checkAuthAction.fulfilled" with thunk "checkAuthAction', async () => {
-    mockAxiosAdapter.onGet(ApiRoute.Login).reply(200);
-
-    await store.dispatch(checkAuthAction());
-    const actions = extractActionsTypes(store.getActions());
-
-    expect(actions).toEqual([
-      checkAuthAction.pending.type,
-      checkAuthAction.fulfilled.type,
-    ]);
-  });
-});
