+++ src/services/thunk/thunk-test/thunk.test.ts
@@ -1,0 +1,88 @@
+import {createApi} from '../../api';
+import MockAdapter from 'axios-mock-adapter';
+import thunk from 'redux-thunk';
+import {Action} from 'redux';
+import {State} from '../../../types/type-store';
+import {ThunkDispatch} from 'redux-thunk';
+import {configureMockStore} from '@jedmao/redux-mock-store';
+import {ApiRoute} from '../../../const';
+import {checkAuthAction} from '../check-auth-action';
+import {offersMock} from '../../../mock/offers/offer-mocks';
+import {fetchOffersAction} from '../fetch-offers';
+
+type AppThunkDispatch = ThunkDispatch<State, ReturnType<typeof createApi>, Action>
+
+const extractActionsTypes = (actions: Action<string>[]) => actions.map(({type}) => type);
+
+describe('Async actions', () => {
+  const axios = createApi();
+  const mockAxiosAdapter = new MockAdapter(axios);
+  const middleware = [thunk.withExtraArgument(axios)];
+  const mockStoreCreator = configureMockStore<State, Action<string>, AppThunkDispatch>(middleware);
+  let store: ReturnType<typeof mockStoreCreator>;
+
+  beforeEach(() => {
+    store = mockStoreCreator({});
+  });
+
+  it('should dispatch "checkAuthAction.pending" and "checkAuthAction.fulfilled" with thunk "checkAuthAction', async () => {
+    mockAxiosAdapter.onGet(ApiRoute.Login).reply(200);
+
+    await store.dispatch(checkAuthAction());
+    const actions = extractActionsTypes(store.getActions());
+
+    expect(actions).toEqual([
+      checkAuthAction.pending.type,
+      checkAuthAction.fulfilled.type,
+    ]);
+  });
+
+
+  it('should dispatch "checkAuthAction.pending" and "checkAuthAction.rejected" when server response 400', async() => {
+    mockAxiosAdapter.onGet(ApiRoute.Login).reply(400);
+
+    await store.dispatch(checkAuthAction());
+    const actions = extractActionsTypes(store.getActions());
+
+    expect(actions).toEqual([
+      checkAuthAction.pending.type,
+      checkAuthAction.rejected.type,
+    ]);
+  });
+
+
+  describe('fetchQuestionAction', () => {
+    it('should dispatch "fetchOffersAction.pending", when server response 200', async() => {
+
+      const mockOffers = offersMock;
+
+      mockAxiosAdapter.onGet(ApiRoute.Offers).reply(200, mockOffers);
+
+      await store.dispatch(fetchOffersAction());
+
+      const emittedActions = store.getActions();
+      const extractedActionsTypes = extractActionsTypes(emittedActions);
+      const fetchQuestionsActionFulfilled = emittedActions.at(1) as ReturnType<typeof fetchOffersAction.fulfilled>;
+
+      expect(extractedActionsTypes).toEqual([
+        fetchOffersAction.pending.type,
+        fetchOffersAction.fulfilled.type,
+      ]);
+
+      expect(fetchQuestionsActionFulfilled.payload)
+        .toEqual(mockOffers);
+    });
+
+    it('should dispatch "fetchOffersAction.pending", "fetchOffersAction.rejected" when server response 400', async () => {
+      mockAxiosAdapter.onGet(ApiRoute.Offers).reply(400, []);
+
+      await store.dispatch(fetchOffersAction());
+      const actions = extractActionsTypes(store.getActions());
+
+      expect(actions).toEqual([
+        fetchOffersAction.pending.type,
+        fetchOffersAction.rejected.type,
+      ]);
+    });
+  });
+});
